Lab-4-Parallelization/requirements.txt
# Core LangChain + Snowflake Integration
langchain-snowflake>=0.2.2
langchain-core>=0.3.0
langgraph>=0.2.0

# Snowflake Connectivity
snowflake-connector-python>=3.0.0
snowflake-snowpark-python>=1.0.0

# External APIs
requests>=2.31.0
googlemaps>=4.10.0

# UI
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Async Support
aiohttp>=3.9.0
